requirements.txt
APScheduler==3.10.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.3
Flask-APScheduler==1.12.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
OpenAIAuth==0.3.2
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
revChatGPT==2.3.3
six==1.16.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
Werkzeug==2.2.3
